Serve welcome page from static files
