+++ public/styles.css
@@ -1,0 +1,165 @@
+@import url('https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css');
+
+body {
+    font-family: Arial, sans-serif;
+    background-color: #121212;
+    color: #e0e0e0;
+    margin: 0;
+    padding: 0;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    min-height: 100vh;
+}
+.container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    padding: 20px;
+    box-sizing: border-box;
+}
+.custom-container {
+    position: relative;
+    text-align: center;
+    background: #1e1e1e;
+    padding: 20px;
+    border-radius: 10px;
+    box-shadow: 0 0 15px rgba(0, 0, 0, 0.5);
+    margin: 10px;
+    width: 100%;
+    max-width: 600px;
+}
+h1 {
+    color: #bb86fc;
+    font-size: 1.5em;
+}
+p {
+    color: #b0bec5;
+    font-size: 1em;
+}
+.online-indicator {
+    position: absolute;
+    top: 10px;
+    right: 10px;
+    width: 10px;
+    height: 10px;
+    background-color: #00e676;
+    border-radius: 50%;
+}
+.footer {
+    margin-top: 20px;
+    font-size: 0.9em;
+}
+.code {
+    background: #333;
+    padding: 2px 5px;
+    border-radius: 3px;
+    color: #76ff03;
+}
+.link {
+    margin-top: 20px;
+    font-size: 0.9em;
+    color: #b0bec5;
+}
+.github-icon {
+    margin-right: 5px;
+    color: #76ff03;
+}
+.link a {
+    color: #76ff03;
+    text-decoration: none;
+}
+.link a:hover {
+    text-decoration: underline;
+}
+.search-bar {
+    width: 100%;
+    padding: 10px;
+    margin-bottom: 20px;
+    border-radius: 5px;
+    border: 1px solid #333;
+    background-color: #333;
+    color: #fff;
+    font-size: 1em;
+}
+.tools-container {
+    border: 1px solid #333;
+    border-radius: 5px;
+    max-height: 300px;
+    overflow-y: auto;
+    margin-top: 10px;
+    padding: 10px;
+}
+.tool-item {
+    display: flex;
+    align-items: center;
+    width: 100%;
+    margin: 10px 0;
+    padding: 10px;
+    border-radius: 5px;
+    background-color: #333;
+    color: #fff;
+    cursor: pointer;
+    font-size: 1em;
+    box-sizing: border-box;
+}
+.tool-icon {
+    flex-shrink: 0;
+    margin-right: 10px;
+    font-size: 1.5em;
+}
+.tool-name {
+    flex: 1;
+    font-weight: bold;
+    text-align: left;
+}
+.tool-description {
+    flex: 2;
+    text-align: left;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+}
+.no-tools {
+    text-align: center;
+    width: 100%;
+}
+.modal {
+    display: none;
+    position: fixed;
+    z-index: 1;
+    left: 0;
+    top: 0;
+    width: 100%;
+    height: 100%;
+    overflow: auto;
+    background-color: rgba(0,0,0,0.4);
+    padding-top: 60px;
+}
+.modal-content {
+    background-color: #1e1e1e;
+    margin: 5% auto;
+    padding: 20px;
+    border: 1px solid #888;
+    width: 90%;
+    max-width: 600px;
+    border-radius: 10px;
+    color: #e0e0e0;
+}
+.modal-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+.close {
+    color: #aaa;
+    font-size: 28px;
+    font-weight: bold;
+    cursor: pointer;
+}
+.close:hover,
+.close:focus {
+    color: #fff;
+    text-decoration: none;
+}
