+++ vercel.json
@@ -1,16 +1,16 @@
 {
-	"functions": {
-		"api/welcome.ts": {
-			"memory": 1024,
-			"maxDuration": 60
-		},
-		"api/toolsHandler.ts": {
-			"memory": 1024,
-			"maxDuration": 60
-		}
-	},
-	"rewrites": [
-		{ "source": "/", "destination": "/api/welcome" },
-		{ "source": "/belt", "destination": "/api/toolsHandler" }
-	]
+        "functions": {
+                "api/welcome.ts": {
+                        "memory": 1024,
+                        "maxDuration": 60
+                },
+                "api/toolsHandler.ts": {
+                        "memory": 1024,
+                        "maxDuration": 60
+                }
+        },
+        "rewrites": [
+                { "source": "/", "destination": "/index.html" },
+                { "source": "/belt", "destination": "/api/toolsHandler" }
+        ]
 }
